+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 * [#3229](https://github.com/xmake-io/xmake/issues/3229): Fix find rc.exe for vs2015
 * [#3271](https://github.com/xmake-io/xmake/issues/3271): Fix macro defines with spaces
 * [#3273](https://github.com/xmake-io/xmake/issues/3273): Fix nim link error
+* [#3286](https://github.com/xmake-io/xmake/issues/3286): Fix compile_commands for clangd
 
 ## v2.7.5
 
@@ -1524,6 +1525,7 @@
 * [#3229](https://github.com/xmake-io/xmake/issues/3229): 修复 vs2015 下找不到 rc.exe 问题
 * [#3271](https://github.com/xmake-io/xmake/issues/3271): 修复支持带有空格的宏定义
 * [#3273](https://github.com/xmake-io/xmake/issues/3273): 修复 nim 链接错误
+* [#3286](https://github.com/xmake-io/xmake/issues/3286): 修复 compile_commands 对 clangd 的支持
 
 ## v2.7.5